Integration test TvDb token retrieval
Set up integration tests that make requests to the TvDb to ensure
requests are formatted properly. Use test implementations of the Emby
interfaces as not all dependencies required by the real implementations
are available.

The TvDb API key is stored in a json file which is not included in git.