Store TvDb episode data as Json
Use the Json serialiser store the TvDb episode data as it can be
configured to handle serialising Maybes more easily, and can handle
deserialising the files via constructors.

It also allows the same type to be used when receiving data from TvDb
as when reading the same data from the file.