Make property mappings more flexible
Because property mappings need to be treated as the same type on the
configuration page, it isn't possible to know what source types are
required when reading them back.

Give the property mappings the ability to indicate whether they accept
a particular source, so available source data objects can be tested
rather than testing for different implementations of IPropertyMapping.