Test dependency injection configuration
Ensure that all registered types can be created.

In order to void circular dependencies relating to PluginConfiguration,
and because the configuration must be serialisable (i.e. no constructor
parameters), remove some dependencies and replace them with method
parameters.

Set up property mapping config on the configuration as part of plugin
creation, as the plugin is resolved and can provide the dependencies.