Rearrange mapping configuration
Rename and reorganise the types relating to property mapping as a lot
of the existing names were not descriptive and too similar.

Avoid referencing types specific to a particular source explicitly.

Combine the mapping configuration for the different metadata types into
a single type for each source to avoid having too many separate classes.