Allow series to be marked as non-anime
To work around incorrect matches (e.g. Future Man), allow the user to
enter a list of series names to be ignored.

If the series already has an AniDb Id, this clears it as well as
skipping series recognition.

The other providers (season and episode) still run, but without the
series Id they never find any matches.