Update Emby version
The latest version of Emby removes IHasMetadata, Emby code now uses
BaseItem instead so do the same.

Support for integrating with the built-in dependency injection is also
removed, so create a separate container that falls back to the
built-in container for Emby types, and set up a type for each entry
point that resolves the implementation from the new container.

This maintains the existing behaviour as the Emby dependencies still
come from the built-in container.