Create a series provider using the new process
Add a series provider which uses IMediaItemProcessor to get its results.

Include the type of the Emby item on the media item type to avoid
explicit casts when getting the ultimate result back.

Update/set provider ids as part of creating the Emby metadata result.

Take account of the exclude list when processing series data.