Add series provider integration test
Set up a test that runs the full process for getting data for a series
from beginning to end (excluding loading files from remote sources).

Ensure that provider ids are set on the returned metadata, and that the
name is always set by the library source to prevent mismatches on
future runs.