Set genres, people, tags, and studios for episodes
This data comes from series objects and so series data has to be loaded
when processing an episode. Add two new source data loaders to handle
this, and give them distinct sources ('Additional' sources) to avoid
series data being returned when data for a particular source is
requested for an episode.